requirements.txt
tkinter
speechrecognition==3.10.0
pyaudio==0.2.11
wave
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
gspread==5.10.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-api-python-client==2.88.0
oauth2client==4.1.3
